🚚 refactor: move database to client
