+++ src/database/client/core/__tests__/db.test.ts
@@ -1,7 +1,7 @@
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 
-import { DBModel } from '@/database/core/types/db';
-import { DB_File } from '@/database/schemas/files';
+import { DBModel } from '@/database/client/core/types/db';
+import { DB_File } from '@/database/client/schemas/files';
 
 import { LocalDB } from '../db';
