+++ src/database/client/models/plugin.ts
@@ -1,6 +1,6 @@
 import { LobeChatPluginManifest } from '@lobehub/chat-plugin-sdk';
 
-import { BaseModel } from '@/database/core';
+import { BaseModel } from '@/database/client/core';
 import { LobeTool } from '@/types/tool';
 import { merge } from '@/utils/merge';
