+++ src/services/__tests__/topic.test.ts
@@ -1,12 +1,12 @@
 import { Mock, beforeAll, beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { CreateTopicParams, TopicModel } from '@/database/models/topic';
+import { CreateTopicParams, TopicModel } from '@/database/client/models/topic';
 import { ChatTopic } from '@/types/topic';
 
 import { topicService } from '../topic';
 
 // Mock the TopicModel
-vi.mock('@/database/models/topic', () => {
+vi.mock('@/database/client/models/topic', () => {
   return {
     TopicModel: {
       create: vi.fn(),
