+++ src/store/chat/slices/plugin/action.ts
@@ -4,9 +4,8 @@ import { Md5 } from 'ts-md5';
 import { StateCreator } from 'zustand/vanilla';
 
 import { PLUGIN_SCHEMA_API_MD5_PREFIX, PLUGIN_SCHEMA_SEPARATOR } from '@/const/plugin';
-import { CreateMessageParams } from '@/database/models/message';
 import { chatService } from '@/services/chat';
-import { messageService } from '@/services/message';
+import { CreateMessageParams, messageService } from '@/services/message';
 import { ChatStore } from '@/store/chat/store';
 import { useToolStore } from '@/store/tool';
 import { pluginSelectors } from '@/store/tool/selectors';
